backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.5
mangum==0.19.0
bson==0.5.10
motor==3.6.1
beanie==1.29.0
python-dotenv==1.0.1
langchain==0.3.14
langchain-community==0.3.14
langgraph==0.2.64
langchain-groq==0.2.3
python-multipart==0.0.20
pinecone==5.4.2
pypdf==5.1.0
python-docx==1.1.2
pillow==11.1.0
pytesseract==0.3.13
